Refactor chart.py into multiple modules
- Move PsychroCurve(s) classes definitions to separate module
- Rename equations to chartdata and move there all logic for curve
data generation
- Remove logger, verbose parameters. Use just logging for errors.
- Remove psychrometric equations:
  * pressure_by_altitude: using GetStandardAtmPressure directly
  * wet_bulb_temperature: only used in tests, use GetTWetBulbFromRelHum
- Adapt tests
